[JENKINS-70309] Restore the selectedJDK parameter
Revert "Added support for REST API buildWithParameters call"

https://issues.jenkins.io/browse/JENKINS-70309 reports that the
selectedJDK parameter was removed from JDK parameter plugin 1.1.  The
commit that removed the capability was merged to the plugin repository
in 2015 with no intervening release that included that commit until the
2022 release of JDK parameter plugin 1.1.

No users of the released plugin could have relied on the feature added
in 22a71621a9f9f52b5f4dfc9e018fd1de3b996f7b because it was not available
until very recently.  Remove that feature in case it solves the issue
https://issues.jenkins.io/browse/JENKINS-70309

This reverts commit 22a71621a9f9f52b5f4dfc9e018fd1de3b996f7b.