Revert "this change allows this plugin to work with the multijob plugin"
Change was never released in a plugin version.  The multijob plugin is
deprecated, so adding support of multijob plugin does not help current
Jenkins users very much.

Change was merged in 2020, never released.  Rather than risk a
regression due to the addition of support for a deprecated plugin, let's
create the plugin without that addition.

This reverts commit a4081131d2955981b629db1ad9a1e93e78231ec5.